add a simple replay script
I'd prefer not to debug the modbus code while actively
connected to the inverter. This provides an alternative - you
can capture traffic to a file, and then this script allows you
to replay that file through the framer and plant update code
offline.

One way to capture the traffic is simply to add a file writer to
the network consumer task, to simply write the raw data to a file.
Another way is to make a passive connection, eg using socat:

 socat -u -x TCP:address:port CREATE:logfile

Then run the app or similar - all replies to that will be
broadcast to this connection too. (This will not respond to
heartbeat requests, and so will get disconnected after a few
minutes.)