Restructure interface between client and plant
This is slightly experimental, big-ish change.  The plant now
takes responsibility for tracking the register set available
on the inverter, and the number of batteries. Each time there's
an incoming ReadRegistersResponse, it looks to see if it
references registers high that it has previously seen. If so,
it updates some tracking attributes.

Then for a client refresh, ask the plant about the highest
registers it has seen.

client.detect_plant() then really just has to send out requests
for high registers. If there are any replies, the plant will
handle that itself.

This also means that, for your own scripts, you can create
the plant with the registers you know you have, and don't need
to detect anything.

When you do a refresh, you should be able to ask for a
subset of the registers, but that's not done yet. Just
a detail in choosing the best interface.

Also extends the ability to subclass Plant and receive
callbacks for register updates. Primarily because holding
register changes are relatively rare, so I don't want to
actively poll for changes. And also because the way the
inverter is implented, it does a lot of work to get just
a single register value. But I might not keep it if I
can simplify the inverter interface to make it more efficient.

Also, change NotImplemented to InvalidFrame for unknown transparent
packets. Since firmware upgrade, I've been getting unknown code #22
from time to time.