strip Pydantic out of the model sub-module.
The code was using Pydantic to generate the Inverter and
Battery classes from the register descriptions. I'm sure
it's all very clever, but massively over-complicated.

My main problem with it is that when you access plant.inverter,
it pre-converts and scales all the known registers into the
human-readable formats, which is wasted work for fields that
don't get accessed. (And as firmwares improve, more and more
registers are added.)

Another feature is that the instance is immutable, which means
subsequent updates to the plant won't be reflected. This may
be desirable, but I'm not convinced.

So replace it all with a simpler scheme: Inverter and Battery
are straightforward, lightweight classes - instantiating them
does nothing up-front. Instead, when you access inverter.field,
it invokes __getattr__('field') to supply the missing attribute,
and only then is the register converted.

This also means that each time you access inverter.field, it
will be converted.

You can also use inverter.get('field') which ends up in exactly
the same place, if you prefer that sort of interface.

I merged register_cache.py into register.py to avoid
circular import problems.