Reimplement the way register write checks are done.
commands.py had its own list of writable registers. Instead,
extend the register definitions in inverter.py to indicate
not only whether they are writable, but also what the valid
range of values is.

Still work in progress, but this is primarily just to
outline the scheme. No current way to set multi-register
values, specifically start-end timeslots, but I think perhaps
simplest just to add register aliases for the single register
halves, which are writable, as well as the composite value,
which is read-only.

A new method in inverter.py will take a register name
and proposed value, and if the value is within range
(which implies the register is writable), return the
register number to be written. Otherwise raise an exception.

A new helper in command.py uses this to prepare a
WriteHoldingRegisterRequest.

Fix a typo in register.py in the fn used for converting
the system time.