Small changes to inverter.name accesses.
I had mistyped the name of a register on an inverter.x access,
and it just quietly returned None. That's no use - if it was a
real attribute, it would be an exception.

Perhaps get() should be renamed, since on a dictionary, that
does quietly ignore key errors. Or I could have a _get()
that the transparent access uses, and then get() can do
try:
   return _get(...)
except KeyError:
    return default

Related to this... if the register cache doesn't have a
value for a (valid) register name, return None rather than 0.
Otherwise caller cannot distinguish a missing value from a 0 value.