Add a commands.__getattr__ to avoid lots of trivial setters.
Rather than adding lots of trivial methods which look like

def set_xyz(value):
    return [ write_named_register('xyz', value) ];

provide a module __getattr__ which does it all transparently.

Automatically generate the module docstring to include a list of setters
made available this way. Unfortunately, I've not yet found a way to do it
lazily when __doc__ is accessed, so it happens unconditionally on module load.