Change imports to use relative paths.
Rather than explicitly specifying  givenergy_modbus. on
every import, use relative naming with . and ..
This means that, while developing, it is possible to locally rename
the directory containing the tree and, with no source changes
required, it will happily run as the new package name.

Also, simplify an import in register_cache.py - it was
conditionally importing TimeSlot at the top of the file,
but then unconditionally importing it later on. So just
do the unconditional import at the top.