Quite a large update to the plant implementation
Use a python set to remember which base registers have been
seen in ReadRegister responses (and therefore which exist on
the connected system) - this can be used as a basis for a plant refresh.

Similarly, have the plant automatically track how many batteries
there are, by validating the contents of ReadRegister responses
as they come in.

Remove plant.detect_batteries() since that's now automatic.

Add empty methods registers_updated() and battery_updated().
User code can subclass Plant in order to hook these calls
and so receive updates. The former is used for both
ReadRegistersResponse and WriteHoldingRegisterResponse

Add handling for some behaviour oddities in the rubikcube
android app:
-  it uses modbus address 0x30
- when looking for batteries, it seems to start enumerating at
  address 0x30 and upwards, which means we get bogus empty
  register sets of 0x30 and 0x31 - decline to add these to
  the register cache since (without special handling) they'd
  end up overwriting battery #0 in cache slot 0x32

In addition, I seem to regularly see battery input registers
from address 1.

Add plant.refresh() which does the work of choosing which registers
to update.

In client, add optional plant parameter to the constructor
to allow user to supply a custom Plant sub-class, and add
a refresh_plant() which uses information from the plant
to construct an appropriate set of requests.