Reimplement (and simplify) the pdu module.
Transparent message handling in particular. There were lots of intermediate
classes making up the various transparent Request/Reply messages, but seemed
like a lot of extra fuss. In particular, the rules for how to encode and
decode the various bits of messages was distributed throughout the code,
making it all quite hard to follow.

Recognise that they're all doing largely the same thing, with just details about
whether they include a count and/or a values field, so switch to a data-driven
rather than code-driven model: all the work is done in the Transparent base
class, using just a couple of class attributes from the sub-classes
to drive the work. NullResponse can also be handled within this scheme.

This gets rid of lots of trivial classes that weren't contributing terribly much.
And replaces quite a lot of code in the various Transparent classes with
just setting a few class attributes.

Change the decoder to produce a tuple rather than a list for register_values.
Tuples are very slightly more efficient (both space and time). But they are
interface-compatible for the most part, so it's fine for existing code to pass lists.
Type hints can use the more generic Sequence rather than explicitly List.

Framer now directly chooses one of (Heartbeat|Transparent)(Request|Response)
to perform decoding, which gets rid of (Client|Server)IncomingMessage layer,
but does leak a tiny bit of encapsulation out of pdu. But I anticipate moving
framer into pdu/ in which case that won't matter.

Also gets rid of the 'expected_response' stuff, which is used by the
client to match a Response to the originating Request. Previously,
the code was actually creating an appropriate TransparentResponse
instance in order to calculate the shape_hash(). But if we simply define that
the "shape_hash" of the Request and expected Response are the same, we can
just use the original request rather than creating the anticipated response
explictly. (Heartbeat still has expected_response() since client uses that
to create the reply.)

Internally, WriteHoldingRegister messages now use base_register
and register_values[], like the others, rather than register and value. Use some
rather ugly code to provide the familiar register + value interface. For now,
at least...

TransparentMessage.__init__() now performs validation, so there's no need for
ensure_valid_state(). (Which does mean lots of changes to the pytest tests,
to construct request/response messages "properly". (I could use defaults
for everything, but that seems wrong just to reduce work in test code.
There isn't really any sensible default value for a base_register, for
example.)

Simplify code to add the crc to the message. Previously it was making
a second encoder and adding the checksummed bits to that, in order to
calculate the checksum on just those fields. But we can do it instead
by making encoder.crc a proper function, and giving it an optional parameter
to crc just a range of bytes. (Possible now that they're all added in
one place, rather than bits here and there.) Also, add an add_16bit_le()
function rather than messing about byte-swapping then adding big-endian.

Note that when encoding a Response message, the CRC will probably be wrong.
We don't know the required algorithm, but fortunately we currently only ever
need to be able to encode requests. But what I have *seems* to produce the
same crc if I re-encode an incoming message, so it may just be a lucky
accident.